address review: null-guard idle counter, rename migrate test, cosmetic
Two warnings + five suggestions from /review 290:

Warnings:
- sched.c idle_task_func: NULL-guard the sched_diag_idle_loops
  deref. On PLATFORM_HAS_NC_MEMORY the array is pointer-backed
  and scheduler_init's ncmem_alloc could theoretically return
  NULL if the 2 MB NC arena is exhausted. scheduler_init would
  panic first in practice, but the check is cheap.
- test_scheduler.c: rename test_migrate_ready_task_moves_queues
  to test_migrate_ready_task_updates_assigned_cpu — the rename
  matches what the test actually asserts now that the racy
  queue-walks are gone. Updated the doc comment to explain what
  IS and ISN'T covered.

Suggestions:
- sched.c sched_try_steal: increment sched_diag_steal_attempts
  BEFORE the isolated-CPU early return, so the `cpu` shell diag
  reports isolated CPUs as "attempted but skipped" rather than
  "never entered steal path".
- exceptions.c: split the PSCI CPU_OFF banner into CPU-0 vs
  secondary messages. CPU 0 fault is "system unable to
  continue"; secondary fault is "offlining, remaining CPUs
  continue." Clearer operator-visible intent.
- test_integration.c: drop the empty setUp definition; the
  explanatory comment is enough, Unity's weak default is fine.
- test_coop_preempt.c: harmonize upper-bound type to uint32_t
  in both tick-count tests (one was uint64_t, sibling was
  uint32_t). Both evaluate identically; consistency only.
- test_scheduler.c test_policy_heuristic_distributes_tasks:
  drop the irq_save/irq_restore pair and its misleading comment.
  IRQ disable on the running CPU doesn't prevent other CPUs
  from stealing; the captured assigned_cpu read inside the
  add-loop is what makes the test stable, not the IRQ mask.

Verified on pi-5-1: 3/3 boot_test runs reach "All test suites
passed!" (100% green).

+++ kernel/tests/test_coop_preempt.c
@@ -64,7 +64,7 @@ static void test_pit_ticks_advances_over_time(void)
      * cpu_count × 5 ticks / 50 ms. Bound scales with cpu_count
      * with a 2× safety margin for yield-loop bursts and the
      * occasional boundary overshoot at the 10 ms period edge. */
-    uint64_t upper = (uint64_t)cpu_count * 10;
+    uint32_t upper = cpu_count * 10;
     if (upper < 20) upper = 20;
     TEST_ASSERT_MESSAGE(delta <= upper,
         "pit_ticks advanced too fast — period math may be wrong");
